requirements.txt
pyyaml
pyzmq
kubernetes
aiohttp
python-telegram-bot
slack-sdk 